requirements.txt
asgiref==3.5.2
Django==3.0
django-tastypie==0.14.4
Pillow==9.2.0
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
